server/requirements.txt
GitPython>=3.1.40
requests>=2.31.0
PyYAML>=6.0.1
Flask>=2.3.0

